services/news_stream/requirements.txt
alpaca-py==0.21.0
psycopg2-binary==2.9.9
websockets==12.0
